Validate compliance reports
This commit sends a message back to the 'uploadvalidation' queue in
Insights. The message validates the payload whether the payload
contained a valid XCCDF Report or not.

Additionally I started storing reports in tmp/storage, as using a Ruby
Tempfile as it was doing before wasn't enough. The reason is that
Tempfiles get destroyed as soon as the garbage collector runs.
Therefore the timeline went like this:

 1. Consumer creates Tempfile
 2. Only the *path* is passed to ActiveJob.
 3. Consumer exits. Tempfile gets thrashed
 4. ActiveJob keeps on running, but the file has been destroyed