Fix GET/POST host from/to inventory
Before this commit, this app was using an old way of querying the
inventory. The API endpoint changed, and so did the body of the request
needed to create a host.

In addition to that, now the Host ID is automatically synced from
inventory, so that when the Host is created, its ID is picked up
automatically from there, instead of creating 2 different IDs.

Aside from that, we have now changed the consumer to fix the validation
format, which was also adhering to an old validation message standard in
insights-upload.