Use file.delete to clear empty files instead of unlink
To save the downloaded file, e use a 'TempFile' object to
ensure it's destroyed at some point if the object reference is no longer
used. For some reason, in ruby the TempFile object is automatically
converted to a File object when you call .close on it. At that
point the code breaks, because you can call tempfile.unlink, but not
file.unlink (it's not a public instance method of File)