RHICOMPL-178 - Optimize Rule results policy fetching
This PR fixes the policy method to retrieve rule_results on
api/compliance/rule_results. Before this PR, the policy
was iterating on all of the rule_results, which was incredibly
slow.

Aside from that, output was not paginated unless 'search' was
passed. Fixing the pagination forced me to change other controllers to
match it, as a side effect all other endpoints pagination are
now fixed.