RHICOMPL-363 - Make uniqueness constraint for profile account_id and name (#249)
Currently if many threads start to upload reports with the same profile
for the first time to an account, it could happen that 6 sidekiq pods
(at the moment these are 6 threads but it could change) start processing
the report at the same time.

Since all data is saved on the database at once, the Rails uniqueness
constraints may NOT apply in this case of concurrent parsing. It will
end with multiple profiles with the same ref_id, name and account_id
which is not what is expected.

This PR introduces a DB constraint so that when this happens, only 1 of
the jobs will actually save the profile, the first one to finish
parsing. The rest of them will likely fail - which is EXPECTED. As the
job is going to retry 3 times, it will retry the parsing and the 2nd
time `Profile.find_or_initialize` will find the pre-existing profile, so
it's eventually consistent.