Find policy for profile (#404)
* Remove cache expiration date (#400)

These values are meant to be expired during report parsing, in the
invalidate_cache method. No need to manually expire them in a week.

* RHICOMPL-586 - Batch profile query on ruleObjectsFailed (#398)

Signed-off-by: Daniel Lobato Garcia <me@daniellobato.me>

* Determine policy from a profile (#399)

Signed-off-by: Andrew Kofink <akofink@redhat.com>

* Add policy to GQL profile type (#402)

This field is meant to return the "Policy" for a profile, so if the
profile is external, this would be any policy in the same benchmark
ref_id, same ref_id, different benchmark version, if it exists, or null.

* Destroy all profiles in a policy on policy destroy (#403)

Signed-off-by: Andrew Kofink <akofink@redhat.com>

* Sync compliance DB with inventory task (#384)

* Sync compliance DB with inventory task

This rake task simply sends requests to the inventory for every host,
and checks whether the host exists in the inventory. If it doesn't
exist, it deletes it from our DB.

Signed-off-by: Daniel Lobato Garcia <me@daniellobato.me>

* Delete hosts async

* Get rid of N+1 on hosts

* Don't delete orphaned profiles (#333)

Signed-off-by: Andrew Kofink <akofink@redhat.com>

* Remove legacy GQL queries not used anymore (#364)

* Remove legacy GQL queries not used anymore

* Update GQL tests

* Update test for rubocop

* Update GQL schema

Co-authored-by: Andrew Kofink <akofink@redhat.com>