RHICOMPL-631 - Update hostnames based on 'update' message
On Apr 28th, there will be a new event to tell apps to update hostnames
if the inventory hostname changed. On the UI, this shouldn't affect us
we already always display the inventory hostname (unless the
host is unsynced for whichever reason and is only on compliance,
this should not happen in prod though).

On the API, this would be visible as we don't contact the
inventory for anything there. So we should update the names
according to the message to serve the right name.

https://platform-docs.cloud.paas.psi.redhat.com/backend/inventory.html#host-patching
for more info on the message format