RHICOMPL-598 - Resolve rule_ids before querying references
Resolving the query for all Rule ids before trying to get all of the
RuleReferencesRules results in a massive performance improvement.

There's a penalty on running subqueries like "IN (SELECT
"rule_results"."rule_id" FROM "rule_results" WHERE
"rule_results"."test_result_id" = $1)" - versus just resolving the call
and passing the array of rule IDs.

In this example, we have already resolved the rule_ids in line 40, on
initialize_rules_context, so using a subquery is suboptimal when the
data is already in memory.