RHICOMPL-622 - Deleting external profiles deletes Profile
Before this change, if you deleted a Report in the UI, external profiles
are still kept in the database. This doesn't make sense, as those
reports are only meant to exist to hold the definition of the profile
for the Report, so if the report goes away, so does the profile and the
profile host associations.

This is visible through the API but also on the UI, after removing an
external Report, the SystemsTable shows the Host still linked to the
external profile.