[stable] Business collector for prometheus_exporter (#455)
* RHICOMPL-645 - Prometheus collector for business metrics (#447)

* RHICOMPL-645 - Prometheus collector for business metrics

Adding a collector for prometheus_exporter. This runs every time
/metrics is called, which is every 60s, from the prometheus_exporter
deployment. Most other metrics are calculated from the data send to
/send-metrics by the other pods, but this sort of data isn't really
related to any pods runtime.

Timeout has been greatly increased just to prevent any failures,
although benchmarks show the `collect` method is really fast.

[3] pry(#<BusinessCollector>)> Benchmark.measure { collect }
=> #<Benchmark::Tms:0x00007f22216bb288
        @cstime=0.0,
	@cutime=0.0,
	@label="",
	@real=0.030418214002565946,
	@stime=0.002605999999999997,
	@total=0.021810000000000218,
	@utime=0.01920400000000022>

In addition to that, I added binding to 0.0.0.0 as the latest prometheus
binds to 127.0.0.1, which makes the data only visible through the pod.
The docker-compose also now contains the necessary variables to access
the Rails database. --verbose helps us see which collectors are loaded
when prometheus_exporter boots.

* Ignore lib directory in codecov

* Updated to get non-canonical profiles

* Ignore coverage in lib for simplecov and codecov (#453)

When running tests locally, the .codecov.yml is not honored because
coverage is using simplecov without the codecov formatter

Signed-off-by: Andrew Kofink <akofink@redhat.com>

* Bump prometheus_exporter to >= 0.5 (#449)

The -b (bind to address) option in prometheus_exporter was introduced in
version 0.5.0

Signed-off-by: Andrew Kofink <akofink@redhat.com>

Co-authored-by: Andrew Kofink <akofink@redhat.com>