RHICOMPL-650 - Delete cached latest_supported_rhel benchmarks
There's a cached field called 'latest_supported_benchmarks' to avoid
making the same SQL call always to show the latest benchmarks in the
Create Policy wizard. Notice this will NOT affect customers unless we
change the LATEST_SUPPORTED_VERSIONS. This was visible on CI because the
import-job failed (killed OOM or something), so I'm going to clear the
field before the import-latest-rhel-supported job starts.