Add remaining DB uniqueness constraints to stable (#468)
* RHICOMPL-363 - Make uniqueness constraint for profile account_id and name (#249)

Currently if many threads start to upload reports with the same profile
for the first time to an account, it could happen that 6 sidekiq pods
(at the moment these are 6 threads but it could change) start processing
the report at the same time.

Since all data is saved on the database at once, the Rails uniqueness
constraints may NOT apply in this case of concurrent parsing. It will
end with multiple profiles with the same ref_id, name and account_id
which is not what is expected.

This PR introduces a DB constraint so that when this happens, only 1 of
the jobs will actually save the profile, the first one to finish
parsing. The rest of them will likely fail - which is EXPECTED. As the
job is going to retry 3 times, it will retry the parsing and the 2nd
time `Profile.find_or_initialize` will find the pre-existing profile, so
it's eventually consistent.

* Add a unique db constraint for benchmarks

Signed-off-by: Andrew Kofink <akofink@redhat.com>

* Revert "RHICOMPL-363 - Make uniqueness constraint for profile account_id and name (#249)" (#359)

This reverts commit 5dda52598cc7163ce8d8920528c25c0cab48c0da.

* Resolve only dup benchmarks, not rules and profiles (#360)

Signed-off-by: Andrew Kofink <akofink@redhat.com>

* Associate parent profiles outside of own benchmark (#357)

* Associate parent profiles outside of own benchmark

Some profiles cannot be associated to parents within its own benchmark,
such as profiles that come from the phony benchmark (which has no
canonical profiles).

* Remove profiles which are duplicated in the account (phony profiles)

* Update parent profile associator

* Fix typo

* Fix typo on empty benchmarks migration

* Add description to find the exact benchmark for phony profiles

* Find parent profile not matching description if needed

* Choose properly parent  without description

* Remove empty benchmarks only on 'up' migration

* Fix off-by-one error when checking for duplicate profiles

* Do not remove profile if attributes are OK already

* Fix rubocop

* Add unique profile DB constraint (#370)

Signed-off-by: Andrew Kofink <akofink@redhat.com>

* Resolve only duplicate benchmarks. (#397)

The test updates give a much higher level of confidence that we're doing
what we want to do with this resolver.

I've also updated the migration for adding a uniqueness constraint so
that it runs just before the profiles and rules uniqueness migrations

Signed-off-by: Andrew Kofink <akofink@redhat.com>

* Add unique constraint on rules table (#396)

Signed-off-by: Andrew Kofink <akofink@redhat.com>

* Add missing indexes and uniqueness validators (#406)

Signed-off-by: Andrew Kofink <akofink@redhat.com>

Co-authored-by: Daniel Lobato García <me@daniellobato.me>