RHICOMPL-631 - Update hostnames based on 'update' message
* Add account_number to Openshift info (#440)

* Add account_number to Openshift info

Currently, the notifications we are getting contain the User object,
which is useless. Instead, we should get the account number to
help us understand better under which circumstances the exception
happened.

* Fix test to get account_number

* Add namespace to service

* Fix rubocop

* RHICOMPL-631 - Update hostnames based on 'update' message (#429)

* RHICOMPL-631 - Update hostnames based on 'update' message

On Apr 28th, there will be a new event to tell apps to update hostnames
if the inventory hostname changed. On the UI, this shouldn't affect us
we already always display the inventory hostname (unless the
host is unsynced for whichever reason and is only on compliance,
this should not happen in prod though).

On the API, this would be visible as we don't contact the
inventory for anything there. So we should update the names
according to the message to serve the right name.

https://platform-docs.cloud.paas.psi.redhat.com/backend/inventory.html#host-patching
for more info on the message format

* Clear jobs

* Validate message format properly

* Avoid silent fails of InventoryHostUpdated job

Co-authored-by: Daniel Lobato García <me@daniellobato.me>