Change development image to RHEL 8 UBI ruby26
In order to run a dev environment that's much more similar to the one in
production, it's probably a good idea to switch to the base image we are
using already in production.

Since the image is meant to be built using s2i, we can call manually the
assemble script in the Dockerfile.

Additionally, this image is tried and tested for podman if we want to
move to it at some point.