[RHICOMPL-662] Migration of old profile data into new policy records (#610)
* [RHICOMPL-662] Create a service for copying profiles into Policies

The service creates new Policy models for Profiles that are internal and
that do not have a Policy.  It copies some of the attributes from the
profile.  It assigns the same host set from the internal profile.

* [RHICOMPL-662] First pass for migration of profile data to policies

This is mostly intended for testing the migration.  Production should
be migrated only when the whole feature is complete.