[RHICOMPL-1051] Lookup policy on profile clone
For Profile.clone_to it first finds the policy of the passed account.
The policy lookup up starts with the internal profile.

The internal profile:
* must have a policy that have the host assigned
* must be part of the account
* must match the ref_id of the (canonical) profile being cloned
* must match the benchmark ref_id of the profile being cloned

The last benchmark condition is there to ensure that the internal profile is
for the same major OS version.

The policy object is then retrieved from the internal profile.

The second step of the clone_to is to find an existing profile on that
policy.  The existing profile has to match the benchmark id.