[RHICOMPL-1053] Return policy profiles on GraphQL
Policy profiles are to be queried on a profile's policy.
All policy profiles are returned including the internal/initial profile.

Recommended way of querying:

  query Profile($anyPolicyProfileId: String!){
      profile(id: $anyPolicyProfileId) {
          policy {
              profiles {
		  id
                  ssgVersion  # to be added
                  rules {
                      id
                  }
              }
          }
      }
  }

Note, that there can be a case when an internal profile would be the same
as its policy (since there is no Policy GraphQL type, yet) and would be
also found in the list of policy profiles.