[RHICOMPL-1035] GraphQL: Pseudo-policy should inherit name and desc
The pseudo-policy returned on the GraphQL for a profile should inherint
name and description from the policy.
The database records of profiles are keeping the original names and
descriptions (from their parent profiles).
This allows to use the `Profile.name` as policy/profile type, and
`Profile.policy.name` as the policy title (user defined).

Note, that REST API returns name and description of profiles from their
policies.  The policy/profile type on the REST API can be retrieved by
requesting canonical profiles based using the `parent_profile_id`.