feat(RHICOMPL-1086): Delete old test results within a policy
Old test results that are within the same policy for a host will now be
removed when new test results are imported for that host. The same will
happen outside of a policy (i.e. with external reports). The old
behavior would not clean up test reports across SSG versions, but this
new version does.

Signed-off-by: Andrew Kofink <akofink@redhat.com>