feat: RHICOMPL-724 Mark test results as (un)supported
For each test result that we save, we will compare the host's OS to the
supported SSG matrix and appropriately mark the test result as supported
or not.

Old test results will be shown as supported (see DB migration default).
Reports associated to a host with valid OS information and a matching SSG
version in the report's profile will be marked as supported.

For hosts without OS information or for mismatched OS/SSG, the test
result is marked as unsupported.

Signed-off-by: Andrew Kofink <akofink@redhat.com>