fix(Profile): RHICOMPL-1077 GraphQL to return policy name and desc
Initial (main) policy profile should return name and the description
from the policy on GraphQL, which are user defined.
Profile and pseudo policy (profile) would return same name and the
description on GraphQL, if they are the same objects.

This will fix the problem of caching on UI, where the same GQL objects
(identified by id) had a different values depending on context.
It should also improve and fix filtering and displaying in UI.