feat(parsing): RHICOMPL-722 Import supported SSG versions
When parsing a report that has a missing benchmark, import the benchmark
if its version is one of the supported SSG versions. This is required
while we still parse benchmarks from upstream because some downstream
benchmark versions do not exist upstream, and we cannot import them with
the import_rhel_supported job.

Signed-off-by: Andrew Kofink <akofink@redhat.com>