fix(data): RHICOMPL-1301 remove incorrect profiles from policies
A bug introduced in RHICOMPL-1295 caused some profiles to be incorrectly
associated to policies. The initial profile on a policy essentially
determines the policy's policy type. If any profile ref_id does not
match this initial profile's ref_id, it does not belong on the policy
and will be removed along with any test results.

Signed-off-by: Andrew Kofink <akofink@redhat.com>