feat(cert-auth): RHICOMPL-1080 Validate cert-auth
For cert-auth, the identity.system.cn must match the system's owner_id
in the system_profile, otherwise the request is unauthorized. In this
implementation, we simply pass cert-auth identity headers through to
inventory, and the response is filtered by identity.system.cn matching
owner_id in the system_profile. If nothing is returned, the identity
header is invalid for that host.

Signed-off-by: Andrew Kofink <akofink@redhat.com>