feat(rake task): RHICOMPL-1285 to cleanup_db
Because we don't have control over when hosts are removed, but we have
other tables that reference hosts, we should periodically remove
dangling records. Our code does not assume these records are removed and
should continue to behave normally with these dangling records present;
however, this task is in the interest of reducing DB bloat.

Signed-off-by: Andrew Kofink <akofink@redhat.com>