feat(sonarqube): RHICOMPL-1259 Downgrade simplecov, add sonar config
Sonarqube does not work with simplecov > 0.17, unfortunately
It also doesn't work correctly even with this downgraded version because
simplecov generates a coverage report with absolute paths, and the file
paths are different in the sonar-scanner container. I had to manually
update the coverage files to get reporting to work. Additionally, the
reported coverage score does not match what Codecov reports, but it's
fairly close.

Signed-off-by: Andrew Kofink <akofink@redhat.com>