feat(Kafka Producers): RHICOMPL-1242 now with SSL
Our kafka producers will now allow SSL with the same config the
consumers use to connect to kafka. This is fairly difficult to test
in devel without setting up kafka with SSL, but I intend to test this
thoroughly in stage. In any case, it should be able to use plaintext
mode with the change of the racecar environment variables.

Signed-off-by: Andrew Kofink <akofink@redhat.com>