chore(tests): RHICOMPL-1482 prevent minitest nil to nil comparisons
When two variables in minitest have `nil` value, an assertion between
is marked as deprecated and it will fail in future minitest versions.
Issues like this usually mark test cases with not properly covered
fixtures as there's no guarantee that the `nil` value is coming from
where it is expected.

As a solution I'm adding some extra attributes on fixtures globally
where it is possible and set them locally where this would break
other tests.