feat(Benchmarks): RHICOMPL-1773 latest supported from existing
Return latest supported Benchmarks, one for each OS major version, from
the existing Benchmark records.  This ensures that if the Supported SSG
configuration changes and the Benchmark is not yet imported, it will
fallback to the previous existing Benchmark.  This is done by comparing
the versions naturally in database for each OS major version.