fix(entrypoint): RHICOMPL-2115 static common clowder config
Adds clowder-common-bash.sh, a common clowder functions for
configuration, as a static file.  Before the file was downloaded from
GitHub and sha256 checked.  It lead to container start problems, e.g. on
clusters with limited networking (FedRamp) or on GitHub outage.

This makes the runnable code all within a container.

This should fix:
/opt/app-root/src/entrypoint.sh: line 18: : No such file or directory